+++ utils/config/__init__.py
@@ -161,6 +161,7 @@ class Conf(BaseConf):
     concurr_num: str = '16'
     clip_sql = curr_os.clip_sql
     kbShowDhb: bool = False
+    active_workflow: str = ''
 
     def __init__(self, path=None, iname=None):
         self.init_conf()