+++ docs/_github/release_notes.md
@@ -1,5 +1,12 @@
 
+## 🎁 Features
+
++ 新增 [h-comic](https://h-comic.com) 站点支持 by @xulingran in #144
++ rvTool 布局更改，并加入新功能 `以图搜索`，参照 [适用性](https://doc.comicguispider.nyc.mn/feat/) 和文档 [rvTool 以图搜索](https://doc.comicguispider.nyc.mn/feat/#_3-%E5%B7%A5%E5%85%B7%E8%A7%86%E7%AA%97) 部分
++ 聚合搜索增加 [from 剪贴板](https://doc.comicguispider.nyc.mn/feat/ags.html#_2-%E5%89%AA%E8%B4%B4%E6%9D%BF) 方式
++ kemono 的 `作者表格 > 本地收藏` 增加作者头像缓存
+
 ## 🐞 Fix
 
-同 `v2.8.1` ， 见[参考](https://doc.comicguispider.nyc.mn/changelog/history)  
-恢复 pypi (之前的更新方式)  
++ 修复内置浏览器读外链时的空白访问
++ 优化应用壳的外层异常反馈，整理绿色包相关，建议重新下载